+++ R/ichimoku.R
@@ -1,11 +1,12 @@
-# Ichimoku ---------------------------------------------------------------------
+# Ichimoku: Visualization Layer ------------------------------------------------
 
 #' ichimoku
 #'
 #' Create an ichimoku object containing the values for all components of the
 #'     Ichimoku Kinko Hyo cloud chart, ready for plotting. The object includes
-#'     the candlesticks as well as the computed cloud lines: Tenkan-sen,
-#'     Kijun-sen, Senkou span A, Senkou span B and Chikou span.
+#'     a date-time index, OHLC pricing data, candle direction, the cloud lines
+#'     Tenkan-sen, Kijun-sen, Senkou span A, Senkou span B and Chikou span, as
+#'     well as values for the cloud top and base.
 #'
 #' @param x a data.frame or other compatible object, which includes xts,
 #'     data.table, tibble, and matrix.
@@ -16,28 +17,27 @@
 #'     changed as using other values would be invalid in the context of
 #'     traditional Ichimoku analysis.
 #' @param ... other arguments to be passed along. For instance, 'holidays' may
-#'     be passed to the \link{tradingDays} function, used by ichimoku when
+#'     be passed to the \code{\link{tradingDays}} function, used by ichimoku when
 #'     calculating the future cloud.
 #'
-#' @return An ichimoku object is returned with S3 classes of 'ichimoku' and
-#'     'data.frame'.
+#' @return An ichimoku object is returned with S3 classes of 'ichimoku', 'xts'
+#'     and 'zoo'.
 #'
-#'     This object contains a date-time index, OHLC pricing data, candlestick
-#'     direction, as well as the computed ichimoku cloud values, with ticker and
-#'     periodicity parameters set as attributes.
+#'     This object contains a date-time index, OHLC pricing data, candle
+#'     direction, the computed ichimoku cloud values, and cloud top and base
+#'     values, with ticker, periods, and periodicity parameters set as
+#'     attributes.
 #'
 #' @details Calling an ichimoku object automatically invokes its print method,
-#'     which in the absence of arguments, will produce a printout of the data to
-#'     the console as well as a static plot of the cloud chart to the graphical
-#'     device.
+#'     which will by default produce a printout of the data to the console as
+#'     well as a static plot of the cloud chart to the graphical device.
 #'
 #'     For further options, including interactive charting, use plot() on the
 #'     returned ichimoku object to pass further arguments for customising
 #'     the chart.
 #'
-#'     Where an ichimoku object is passed to ichimoku(), the original object
-#'     will be returned (with the ticker amended if the corresponding parameter
-#'     is set).
+#'     Where an ichimoku object is passed to ichimoku(), the ichimoku object is
+#'     re-calculated using the OHLC pricing data contained within.
 #'
 #' @section Further Details:
 #'     Please refer to the reference vignette by running:
@@ -56,9 +56,11 @@ ichimoku <- function(x, ...) UseMethod("ichimoku")
 #' @method ichimoku ichimoku
 #' @export
 #'
-ichimoku.ichimoku <- function(x, ticker, ...) {
-  if(!missing(ticker)) attr(x, "ticker") <- ticker
-  print(x, ...)
+ichimoku.ichimoku <- function(x, ticker, periods = c(9, 26, 52), ...) {
+  if(missing(ticker)) ticker <- attr(x, "ticker")
+  x <- as.data.frame(x)
+  x <- x[!is.na(x$close),]
+  ichimoku.data.frame(x, ticker = ticker, periods = periods, ...)
 }
 
 #' @rdname ichimoku
@@ -89,11 +91,11 @@ ichimoku.data.frame <- function(x, ticker, periods = c(9, 26, 52), ...) {
 
   if(missing(ticker)) ticker <- deparse(substitute(x))
 
-  if(!is.null(rownames(x)) &
+  if(!is.null(rownames(x)) &&
      !inherits(try(as.POSIXct(rownames(x)), silent = TRUE), "try-error")) {
-    date <- as.POSIXct(rownames(x))
-  } else if (length(grep("Index|Date|Time", colnames(x), ignore.case = TRUE)) == 1) {
-    tryCatch(date <- as.POSIXct(x[[grep("Index|Date|Time", colnames(x), ignore.case = TRUE)]]),
+    index <- as.POSIXct(rownames(x))
+  } else if (length(grep("index|date|time", colnames(x), ignore.case = TRUE)) == 1) {
+    tryCatch(index <- as.POSIXct(x[ ,grep("index|date|time", colnames(x), ignore.case = TRUE)]),
       error = function(e) {
         stop("ichimoku: dataset index not convertible to a POSIXct date-time format",
              call. = FALSE)
@@ -103,9 +105,9 @@ ichimoku.data.frame <- function(x, ticker, periods = c(9, 26, 52), ...) {
   }
 
   tryCatch(
-  stopifnot(length(grep("High", colnames(x), ignore.case = TRUE)) == 1,
-            length(grep("Low", colnames(x), ignore.case = TRUE)) == 1,
-            length(grep("Close", colnames(x), ignore.case = TRUE)) == 1),
+  stopifnot(length(grep("high", colnames(x), ignore.case = TRUE)) == 1,
+            length(grep("low", colnames(x), ignore.case = TRUE)) == 1,
+            length(grep("close", colnames(x), ignore.case = TRUE)) == 1),
   error = function (e) {
   stop("ichimoku: clearly-defined High/Low/Close columns not found within the dataset",
        call. = FALSE)
@@ -116,87 +118,91 @@ ichimoku.data.frame <- function(x, ticker, periods = c(9, 26, 52), ...) {
     error = function(e) {
       warning("ichimoku: specified cloud periods invalid - using defaults c(9, 26, 52) instead",
               call. = FALSE)
-      c(9, 26, 52)
+      c(9L, 26L, 52L)
       })
   if(!is.null(tryPeriods)) periods <- tryPeriods
-  p1 <- as.integer(periods[1])
-  p2 <- as.integer(periods[2])
-  p3 <- as.integer(periods[3])
-  tryCatch(stopifnot(p2 < dim(x)[1L]), error = function(e) {
+  p1 <- as.integer(periods[1L])
+  p2 <- as.integer(periods[2L])
+  p3 <- as.integer(periods[3L])
+  xlen <- dim(x)[1L]
+  tryCatch(stopifnot(p2 < xlen), error = function(e) {
     stop("ichimoku: medium cloud period '", p2, "' must be within the length of the dataset",
          call. = FALSE)
     })
 
-  high <- x[[grep("High", colnames(x), ignore.case = TRUE)]]
-  low <- x[[grep("Low", colnames(x), ignore.case = TRUE)]]
-  close <- x[[grep("Close", colnames(x), ignore.case = TRUE)]]
-  if(length(grep("Open", colnames(x), ignore.case = TRUE)) == 1) {
-    open <- x[[grep("Open", colnames(x), ignore.case = TRUE)]]
+  high <- x[ ,grep("high", colnames(x), ignore.case = TRUE)]
+  low <- x[ ,grep("low", colnames(x), ignore.case = TRUE)]
+  close <- x[ ,grep("close", colnames(x), ignore.case = TRUE)]
+  if(length(grep("open", colnames(x), ignore.case = TRUE)) == 1) {
+    open <- x[ ,grep("open", colnames(x), ignore.case = TRUE)]
   } else {
-    open <- c(NA, close[1:(length(close) - 1)])
+    open <- c(NA, close[1:(xlen - 1)])
     warning("ichimoku: opening prices not found - using previous closing prices as substitute",
-            "\n this affects the candles but not the calculation of the cloud", call. = FALSE)
+            "\nThis affects the candles but not the calculation of the cloud", call. = FALSE)
   }
-  candle <- rep("flat", length(close))
-  candle[open > close] <- "down"
-  candle[open < close] <- "up"
 
+  cd <- rep(0, xlen)
+  cd[open < close] <- 1
+  cd[open > close] <- -1
   tenkan <- (maxOver(high, p1) + minOver(low, p1)) / 2
   kijun <- (maxOver(high, p2) + minOver(low, p2)) / 2
   senkouA <- (tenkan + kijun) / 2
   senkouB <- (maxOver(high, p3) + minOver(low, p3)) / 2
-  chikou <- c(close[(p2 + 1):length(close)], rep(NA, p2))
+  chikou <- c(close[(p2 + 1):xlen], rep(NA, p2))
+  cloudTop <- pmax.int(senkouA, senkouB)
+  cloudBase <- pmin.int(senkouA, senkouB)
 
-  periodicity <- min(difftime(date[2], date[1]), difftime(date[3], date[2]),
-                     difftime(date[4], date[3]))
-  extra <- seq.POSIXt(from = date[length(date)], by = periodicity, length.out = p2 + p2)[-1]
+  periodicity <- min(diff(index[1:4]))
+  extra <- seq.POSIXt(from = index[length(index)], by = periodicity, length.out = p2 + p2)[-1]
   if(attr(periodicity, "units") == "days") {
     extra <- extra[tradingDays(extra, ...)][1:p2]
   } else {
     extra <- extra[1:p2]
   }
 
-  cloud <- data.frame(
-    date = c(date, extra),
+  cloud <- xts(cbind(
     open = c(open, rep(NA, p2)),
     high = c(high, rep(NA, p2)),
     low = c(low, rep(NA, p2)),
     close = c(close, rep(NA, p2)),
-    candle = factor(c(candle, rep(NA, p2)), levels = c("down", "flat", "up")),
+    cd = c(cd, rep(NA, p2)),
     tenkan = c(tenkan, rep(NA, p2)),
     kijun = c(kijun, rep(NA, p2)),
     senkouA = c(rep(NA, p2), senkouA),
     senkouB = c(rep(NA, p2), senkouB),
-    chikou = c(chikou, rep(NA, p2))
-  )
+    chikou = c(chikou, rep(NA, p2)),
+    cloudTop = c(rep(NA, p2), cloudTop),
+    cloudBase = c(rep(NA, p2), cloudBase)
+  ), order.by = c(index, extra))
+  attr(cloud, "periods") <- c(p1, p2, p3)
   attr(cloud, "periodicity") <- as.numeric(periodicity, units = "secs")
   attr(cloud, "ticker") <- ticker
 
-  structure(cloud, class = c("ichimoku", "data.frame"))
+  structure(cloud, class = c("ichimoku", "xts", "zoo"))
 }
 
 #' @rdname ichimoku
 #' @method ichimoku default
 #' @export
 #'
-ichimoku.default <- function(x = NULL, ticker, ...) {
+ichimoku.default <- function(x = NULL, ticker, periods = c(9, 26, 52), ...) {
   tryExists <- try(exists(x), silent = TRUE)
   if(inherits(tryExists, "try-error")) {
     message("ichimoku: cannot create an ichimoku object from a '", class(x), "' object")
   } else {
     if(!tryExists) message("Error in ichimoku(", x, "): object '", x, "' not found")
-    else if(missing(ticker)) ichimoku(get(x), ticker = x)
-    else ichimoku(get(x), ticker = ticker)
+    else if(missing(ticker)) ichimoku(get(x), ticker = x, periods = periods, ...)
+    else ichimoku(get(x), ticker = ticker, periods = periods, ...)
   }
 }
 
-#' print.ichimoku
+#' Print Ichimoku Objects
 #'
 #' Custom print method for ichimoku objects.
 #'
 #' @param x an object of class 'ichimoku'.
-#' @param plot set to TRUE by default. Set to FALSE to prevent automatic
-#'     plotting of the ichimoku cloud chart.
+#' @param plot defaults to TRUE. Set to FALSE to prevent automatic plotting of
+#'     the ichimoku cloud chart.
 #' @param ... other arguments to be passed along.
 #'
 #' @return The ichimoku object 'x' passed as parameter.
@@ -212,14 +218,14 @@ ichimoku.default <- function(x = NULL, ticker, ...) {
 #' @examples
 #' cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
 #'
-#' print(cloud, max = 110, row.names = FALSE, digits = 4)
-#' print(cloud[100:110,], plot = FALSE, row.names = FALSE, digits = 4)
+#' print(cloud, max = 110, digits = 4)
+#' print(cloud[100:110,], plot = FALSE, digits = 4)
 #'
 #' @method print ichimoku
 #' @export
 #'
 print.ichimoku <- function(x, plot = TRUE, ...) {
-  if(plot == TRUE) plot.ichimoku(x, ...)
+  if(isTRUE(plot) && (dim(x)[2L] == 12 || dim(x)[2L] == 19)) plot.ichimoku(x, ...)
   NextMethod(print)
   invisible(x)
 }
@@ -235,16 +241,19 @@ NULL
 #' Plot static Ichimoku Kinko Hyo cloud charts from ichimoku objects.
 #'
 #' @param object an object of class 'ichimoku'.
-#' @param from (optional) a starting date/time to subset the plot.
-#' @param to (optional) an ending date/time to subset the plot.
+#' @param window (optional) a date-time window to subset the plot, in ISO-8601
+#'     compatible range strings of the format used for 'xts' objects, for example
+#'     "2020-02-15/2020-08-15" or "2020-02-15/", "/2020-08" or "2020-07".
 #' @param ticker (optional) specify a ticker (or other text) to include in the
 #'     chart heading. If not set, the ticker saved within the ichimoku object
 #'     will be used.
-#' @param theme with a default of 'default'. This can also be set to 'dark',
-#'     'solarized', 'classic' or 'mono' to select the desired colour scheme.
-#' @param gaps set to FALSE by default to remove weekend and holiday gaps. Set
-#'     to TRUE for a continuous timescale axis, but with gaps for non-trading
-#'     days.
+#' @param theme defaults to 'original'. This can also be set to 'dark',
+#'     'solarized' or 'mono'.
+#' @param gaps defaults to FALSE to remove weekend and holiday gaps. Set to TRUE
+#'     for a continuous timescale axis, but with gaps for non-trading days.
+#' @param strat set to TRUE by default. If the ichimoku object contains a
+#'     strategy, the periods for which the strategy results in a position will
+#'     be shaded. Set to FALSE to turn off this behaviour.
 #' @param ... other arguments to be passed along.
 #'
 #' @return Returns a ggplot2 object with S3 classes 'gg' and 'ggplot'.
@@ -261,70 +270,77 @@ NULL
 #' cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
 #'
 #' autoplot(cloud)
-#' autoplot(cloud, from = "2020-05-15", to = "2020-10-30", theme = "dark")
+#' autoplot(cloud, window = "2020-05-15/2020-10-30", theme = "dark")
 #' autoplot(cloud, ticker = "TKR Co.", theme = "solarized", gaps = TRUE)
 #'
 #' @rdname autoplot.ichimoku
 #' @method autoplot ichimoku
 #' @export
 #'
-autoplot.ichimoku <- function(object, from, to, ticker, theme = "default", gaps = FALSE, ...) {
+autoplot.ichimoku <- function(object, window, ticker,
+                              theme = c("original", "dark", "solarized", "mono"),
+                              gaps = FALSE, strat = TRUE, ...) {
 
   periodicity <- attr(object, "periodicity")
   if(missing(ticker)) ticker <- attr(object, "ticker")
-  if(!missing(from)) object <- object[object$date >= from,]
-  if(!missing(to)) object <- object[object$date <= to,]
-  if(theme %in% colnames(ichimoku_themes)) {
-    pal <- ichimoku_themes[, grep(theme, colnames(ichimoku_themes))]
-  } else {
-    pal <- ichimoku_themes[, "default"]
-    message("ichimoku: theme '", theme, "' not found - using 'default' theme instead",
-            "\n available themes are 'default', 'dark', 'solarized' and 'classic'")
-  }
+  if(!missing(window)) object <- object[window]
+  index <- index(object)
+  theme <- tryCatch(match.arg(theme), error = function(e) {
+    message("ichimoku: theme '", theme, "' not found - using 'original' theme instead")
+    "original"
+  })
+  pal <- ichimoku_themes[, theme]
+  candle <- as.factor(object$cd)
+
+  ichimoku_layers_strat <- list(
+    if(isTRUE(strat) && hasStrat(object)) {
+      geom_rect(aes(xmin = .data$posn * (seq_along(index) - 0.5),
+                    xmax = .data$posn * (seq_along(index) + 0.5),
+                    ymin = -Inf, ymax = Inf, na.rm = TRUE), fill = pal[1], alpha = 0.2)
+    })
 
   ichimoku_layers_base <- list(
-    if(!all(is.na(object$senkouB))) {
-    geom_ribbon(aes(ymax = .data$senkouA, ymin = .data$senkouB),
-                fill = pal[1], alpha = 0.6, na.rm = TRUE)},
-    geom_line(aes(y = .data$senkouA), col = pal[2], alpha = 0.6, na.rm = TRUE),
-    geom_line(aes(y = .data$senkouB), col = pal[3], alpha = 0.6, na.rm = TRUE),
-    geom_line(aes(y = .data$tenkan), col = pal[4], na.rm = TRUE),
-    geom_line(aes(y = .data$kijun), col = pal[5], na.rm = TRUE),
-    geom_line(aes(y = .data$chikou), col = pal[6], na.rm = TRUE),
-    geom_linerange(aes(ymax = .data$high, ymin = .data$low, colour = .data$candle),
+    if(!all(is.na(object$cloudTop))) {
+    geom_ribbon(aes(ymax = .data$cloudTop, ymin = .data$cloudBase),
+                fill = pal[1L], alpha = 0.6, na.rm = TRUE)},
+    geom_line(aes(y = .data$senkouA), col = pal[2L], alpha = 0.6, na.rm = TRUE),
+    geom_line(aes(y = .data$senkouB), col = pal[3L], alpha = 0.6, na.rm = TRUE),
+    geom_line(aes(y = .data$tenkan), col = pal[4L], na.rm = TRUE),
+    geom_line(aes(y = .data$kijun), col = pal[5L], na.rm = TRUE),
+    geom_line(aes(y = .data$chikou), col = pal[6L], na.rm = TRUE),
+    geom_linerange(aes(ymax = .data$high, ymin = .data$low, colour = candle),
                    size = 0.3, na.rm = TRUE)
   )
+
   ichimoku_layers_front <- list(
     scale_y_continuous(breaks = function(x) pretty(x, n = 9L)),
-    scale_color_manual(values = c("up" = pal[7], "down" = pal[8], "flat" = pal[9])),
-    scale_fill_manual(values = c("up" = pal[10], "down" = pal[11], "flat" = pal[12])),
+    scale_color_manual(values = c(`1` = pal[7L], `-1` = pal[8L], `0` = pal[9L])),
+    scale_fill_manual(values = c(`1` = pal[10L], `-1` = pal[11L], `0` = pal[12L])),
     labs(x = "Date | Time", y = "Price", title = paste("Ichimoku Kinko Hyo : :", ticker)),
     theme_light(),
-    theme(legend.position = "none")
-  )
-  ichimoku_dark_theme <- list(
-    theme(
-      plot.title = element_text(colour = "#eee8d5"),
-      plot.background = element_rect(fill = "#586e75", colour = NA),
-      panel.background = element_rect(fill = "#002b36", colour = NA),
-      panel.grid = element_line(colour = "#073642"),
-      axis.title = element_text(colour = "#eee8d5"),
-      axis.text = element_text(colour = "#eee8d5"),
-      axis.ticks = element_line(colour = "#eee8d5")
-    )
+    theme(legend.position = "none"),
+    if(theme == "dark") {
+      theme(plot.title = element_text(colour = "#eee8d5"),
+            plot.background = element_rect(fill = "#586e75", colour = NA),
+            panel.background = element_rect(fill = "#002b36", colour = NA),
+            panel.grid = element_line(colour = "#073642"),
+            axis.title = element_text(colour = "#eee8d5"),
+            axis.text = element_text(colour = "#eee8d5"),
+            axis.ticks = element_line(colour = "#eee8d5"))
+    }
   )
 
-  if(gaps == TRUE) {
+  if(isTRUE(gaps)) {
     w <- periodicity * 0.4
-    chart <- ggplot(data = object, aes(x = .data$date)) +
+    chart <- ggplot(data = object, aes(x = index)) +
       ichimoku_layers_base +
-      geom_rect(aes(xmin = .data$date - w,
-                    xmax = .data$date + w,
+      geom_rect(aes(xmin = index - w,
+                    xmax = index + w,
                     ymin = pmin.int(.data$open, .data$close),
                     ymax = pmax.int(.data$open, .data$close),
-                    colour = .data$candle, fill = .data$candle,
+                    colour = candle, fill = candle,
                     open = .data$open, close = .data$close,
-                    date = .data$date), size = 0.3, na.rm = TRUE) +
+                    date = index), size = 0.3, na.rm = TRUE) +
       scale_x_datetime(breaks = function(x) pretty(x, n = 9L),
                        labels = function(x) {
                          if(periodicity < 80000) {
@@ -334,53 +350,40 @@ autoplot.ichimoku <- function(object, from, to, ticker, theme = "default", gaps 
                            }
                          }) +
       ichimoku_layers_front
-
   } else {
-    w <- 0.4
-    chart <- ggplot(data = object, aes(x = seq_along(.data$date), date = .data$date)) +
+    chart <- ggplot(data = object, aes(x = seq_along(index), date = index)) +
+      ichimoku_layers_strat +
       ichimoku_layers_base +
-      geom_rect(aes(xmin = seq_along(.data$date) - w,
-                    xmax = seq_along(.data$date) + w,
+      geom_rect(aes(xmin = seq_along(index) - 0.4,
+                    xmax = seq_along(index) + 0.4,
                     ymin = pmin.int(.data$open, .data$close),
                     ymax = pmax.int(.data$open, .data$close),
-                    colour = .data$candle, fill = .data$candle,
+                    colour = candle, fill = candle,
                     open = .data$open, close = .data$close),
                 size = 0.3, na.rm = TRUE) +
       scale_x_continuous(breaks = function(x) {
-        x <- pretty(seq_along(object$date), n = 9L) + 1
+        x <- pretty(seq_along(index), n = 9L) + 1
         if(x[length(x)] > dim(object)[1L]) x <- x[-length(x)]
         x
       }, labels = function(x) {
         if(periodicity < 80000) {
-          labs <- format(object$date[x], paste("%H:%M", "%d-%b", "%Y", sep = "\n"))
+          labs <- format(index[x], paste("%H:%M", "%d-%b", "%Y", sep = "\n"))
         } else {
-          labs <- format(object$date[x], paste("%d-%b", "%Y", sep = "\n"))
+          labs <- format(index[x], paste("%d-%b", "%Y", sep = "\n"))
         }
      }) +
       ichimoku_layers_front
   }
 
-  if(theme == "dark") chart <- chart + ichimoku_dark_theme
-
   chart
 }
 
-#' iplot interactive ichimoku plot
+#' iplot Interactive Ichimoku Plot
 #'
 #' Plot interactive Ichimoku Kinko Hyo cloud charts from ichimoku objects.
 #'
 #' @param x an object of class 'ichimoku'.
-#' @param from (optional) a starting date/time to subset the plot.
-#' @param to (optional) an ending date/time to subset the plot.
-#' @param ticker (optional) specify a ticker (or other text) to include in the
-#'     chart heading. If not set, the ticker saved within the ichimoku object
-#'     will be used.
-#' @param theme with a default of 'default'. This can also be set to 'dark',
-#'     'solarized', 'classic' or 'mono' to select the desired colour scheme.
-#' @param gaps set to FALSE by default to remove weekend and holiday gaps. Set
-#'     to TRUE for a continuous timescale axis, but with gaps for non-trading
-#'     days.
-#' @param ... other arguments to be passed along.
+#' @inheritParams autoplot
 #'
 #' @return Returns a plotly object with classes 'plotly' and 'htmlwidget'.
 #'
@@ -393,58 +396,45 @@ autoplot.ichimoku <- function(object, from, to, ticker, theme = "default", gaps 
 #' cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
 #'
 #' iplot(cloud)
-#' iplot(cloud, from = "2020-05-15", to = "2020-10-30", theme = "dark")
+#' iplot(cloud, window = "2020-05-15/2020-10-30", theme = "dark")
 #' iplot(cloud, ticker = "TKR Co.", theme = "solarized", gaps = TRUE)
 #' }
 #'
 #' @export
 #'
-iplot <- function(x, from, to, ticker, theme = "default", gaps = FALSE, ...) {
+iplot <- function(x, window, ticker, theme = c("original", "dark", "solarized", "mono"),
+                  gaps = FALSE, ...) {
 
   if(is.ichimoku(x)) {
-    if (requireNamespace("plotly", quietly = TRUE)) {
-      if (gaps == TRUE) {
+    if(requireNamespace("plotly", quietly = TRUE)) {
+      if(isTRUE(gaps)) {
         suppressWarnings(plotly::ggplotly(
-          autoplot(x, from = from, to = to, ticker = ticker, theme = theme, gaps = gaps, ...),
+          autoplot(x, window = window, ticker = ticker, theme = theme, gaps = gaps, ...),
           source = "select", tooltip = c("date", "x", "y", "ymax", "ymin", "open", "close")))
         } else {
           suppressWarnings(plotly::ggplotly(
-            autoplot(x, from = from, to = to, ticker = ticker, theme = theme, gaps = gaps, ...),
+            autoplot(x, window = window, ticker = ticker, theme = theme, gaps = gaps, ...),
             source = "select", tooltip = c("date", "y", "ymax", "ymin", "open", "close")))
           }
       } else {
         message("ichimoku: please install the 'plotly' package for interactive charting")
-        suppressWarnings(print(autoplot(x, from = from, to = to, ticker = ticker, theme = theme, gaps = gaps, ...)
+        suppressWarnings(print(autoplot(x, window = window, ticker = ticker,
+                                        theme = theme, gaps = gaps, ...)
         ))
       }
   } else {
-    message("ichimoku: iplot only works with ichimoku objects - '",
-            deparse(substitute(x)),"' is an object of class ", class(x))
+    message("ichimoku: iplot only works with ichimoku objects")
   }
 }
 
-#' plot.ichimoku
+#' Plot Ichimoku Cloud Chart
 #'
-#' Plot static or interactive Ichimoku Kinko Hyo cloud charts from ichimoku objects.
+#' Plot Ichimoku Kinko Hyo cloud charts from ichimoku objects.
 #'
 #' @param x an object of class 'ichimoku'.
-#' @param i interactive parametmer set to FALSE by default. Set to TRUE to plot
-#'     an interactive rather than a static chart.
-#' @param from (optional) a starting date/time to subset the plot.
-#' @param to (optional) an ending date/time to subset the plot.
-#' @param ticker (optional) specify a ticker, or other text, to include
-#'     in the chart heading. If not set, the ticker saved within the ichimoku
-#'     object will be used.
-#' @param theme with a default of 'default'. This can also be set to 'dark',
-#'     'solarized', 'classic' or 'mono' to select the desired colour scheme.
-#' @param gaps set to FALSE by default to remove weekend and holiday gaps. Set
-#'     to TRUE for a continuous timescale axis, but with gaps for non-trading
-#'     days.
-#' @param ... other arguments to be passed along.
+#' @inheritParams autoplot
 #'
-#' @return Returns either a ggplot2 object with classes 'gg' and 'ggplot', or a
-#'     plotly object with classes 'plotly' and 'htmlwidget' depending on the
-#'     parameters set.
+#' @return Returns a ggplot2 object with classes 'gg' and 'ggplot'.
 #'
 #' @details This function is an S3 method for the generic function plot() for
 #'     class 'ichimoku'. It can be invoked by calling plot(x) on an object 'x'
@@ -458,23 +448,18 @@ iplot <- function(x, from, to, ticker, theme = "default", gaps = FALSE, ...) {
 #' cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
 #'
 #' plot(cloud)
-#' plot(cloud, from = "2020-05-15", to = "2020-10-30", theme = "dark")
-#' \donttest{
-#' plot(cloud, i = TRUE, ticker = "TKR Co.", theme = "solarized", gaps = TRUE)
-#' }
+#' plot(cloud, window = "2020-05-15/2020-10-30", theme = "mono")
 #'
 #' @method plot ichimoku
 #' @export
 #'
-plot.ichimoku <- function(x, i = FALSE, from, to, ticker, theme = "default", gaps = FALSE, ...) {
+plot.ichimoku <- function(x, window, ticker,
+                          theme = c("original", "dark", "solarized", "mono"),
+                          gaps = FALSE, strat = TRUE, ...) {
 
-  if(i == TRUE) {
-    iplot(x, from = from, to = to, ticker = ticker, theme = theme, gaps = gaps, ...)
-  } else {
-    suppressWarnings(print(
-      autoplot(x, from = from, to = to, ticker = ticker, theme = theme, gaps = gaps, ...)
+  suppressWarnings(print(
+    autoplot(x, window = window, ticker = ticker, theme = theme, gaps = gaps, strat = strat, ...)
     ))
-  }
 }
 
 #' is.ichimoku